requirements.txt
tensorflow>=2.10
numpy
pandas
scikit-learn
matplotlib
pyyaml